pyproject.toml
[project]
name = "streamlit-agent-mvp"
version = "0.1.0"
description = "A minimal Streamlit Agent with planning, reasoning, and tool-use capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.25.0",
    "openai>=1.82.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "agent", "parser"]